create one standalone python module
