+++ python/mlc_chat/chat_module.py
@@ -1,4 +1,4 @@
-"""Chat module for MLC chat."""
+"""Chat module for MLC chat in a standalone file, including image module for multimodal-purposes."""
 #! pylint: disable=unused-import, invalid-name
 import tvm
 import tvm._ffi.base
@@ -17,8 +17,6 @@ class ChatModule:
         device_id : int
             The device id.
         """
-        fcreate = tvm.get_global_func("mlc.llm_chat_create")
-        assert fcreate is not None
         if target == "cuda":
             self.device = tvm.cuda(device_id)
         elif target == "metal":
@@ -28,8 +26,15 @@ class ChatModule:
         else:
             raise ValueError("device type not supported yet")
         device_type = self.device.device_type
-        chat_mod = fcreate(device_type, device_id)
 
+        fcreate_chat_mod = tvm.get_global_func("mlc.llm_chat_create")
+        assert fcreate_chat_mod is not None
+        chat_mod = fcreate_chat_mod(device_type, device_id)
+        fcreate_image_mod = tvm.get_global_func("mlc.llm_image_module_create")
+        assert fcreate_image_mod is not None
+        image_mod = fcreate_image_mod(device_type, device_id)
+
+        # chat module related functions
         self.reload_func = chat_mod["reload"]
         self.prefill_func = chat_mod["prefill"]
         self.embed_func = chat_mod["embed"]
@@ -44,6 +49,12 @@ class ChatModule:
         self.evaluate_func = chat_mod["evaluate"]
         self.get_role0 = chat_mod["get_role0"]
         self.get_role1 = chat_mod["get_role1"]
+        # image module related functions
+        self.image_reload_func = image_mod["reload"]
+        self.image_embed_func = image_mod["embed"]
+        self.image_reset_func = image_mod["reset_image_module"]
+        self.image_runtime_stats_text_func = image_mod["runtime_stats_text"]
+        self.image_reset_runtime_stats_func = image_mod["reset_runtime_stats"]
 
     def reload(self, lib: str, model_path: str, app_config_json: str = ""):
         r"""Reload the chat module from the given library and model path.
@@ -169,3 +180,52 @@ class ChatModule:
 
     def evaluate(self):
         self.evaluate_func()
+
+    def reload_image_module(self, lib: str, model_path: str):
+        r"""Reload the image module from the given library and model path.
+
+        Parameters
+        ----------
+        lib : str
+            The library path.
+        model_path : str
+            The model path.
+        """
+        self.reload_func(lib, model_path)
+
+    def reset_image_module(self):
+        r"""Reset the image module, clear its performance record.
+
+        Note
+        ----
+        The model remains the same after :func:`reset_image_module`.
+        To reload module, please use :func:`reload` instead.
+        """
+        self.reset_image_module_func()
+
+    def embed_image_module(
+        self,
+        image: tvm.runtime.NDArray,
+    ):
+        r"""Given an image of type NDArray, get the embedding of the image.
+
+        Parameters
+        ----------
+        image : tvm.runtime.NDArray
+            The user uploaded image.
+        """
+        return self.embed_func(image)
+
+    def runtime_stats_text_image_module(self) -> str:
+        r"""Get the runtime stats text (image encoding speed).
+
+        Returns
+        -------
+        stats : str
+            The runtime stats text.
+        """
+        return self.runtime_stats_text_func()
+
+    def reset_runtime_stats_image_module(self):
+        r"""Reset the runtime stats."""
+        self.reset_runtime_stats_func()
